renamed txtl.py to extracts.py to avoid confusion. Added outline for BasicBuffer class, cleaned up other code
